+++ src/components/Restaurant/index.scss
